feat: configure page and form from CRM UI
(cherry picked from commit cd366839fc65319e736b6e126d9ebc3220f70d35)
